add initial liq
