+++ dex/pair/src/pair_actions/mod.rs
@@ -1,0 +1,3 @@
+pub mod common_methods;
+pub mod common_result_types;
+pub mod initial_liq;
