+++ dex/pair/wasm-pair-full/src/lib.rs
@@ -21,7 +21,6 @@ multiversx_sc_wasm_adapter::endpoints! {
     (
         init => init
         upgrade => upgrade
-        addInitialLiquidity => add_initial_liquidity
         addLiquidity => add_liquidity
         removeLiquidity => remove_liquidity
         removeLiquidityAndBuyBackAndBurnToken => remove_liquidity_and_burn_token
@@ -75,6 +74,7 @@ multiversx_sc_wasm_adapter::endpoints! {
         pause => pause
         resume => resume
         getState => state
+        addInitialLiquidity => add_initial_liquidity
         getLpTokensSafePriceByDefaultOffset => get_lp_tokens_safe_price_by_default_offset
         getLpTokensSafePriceByRoundOffset => get_lp_tokens_safe_price_by_round_offset
         getLpTokensSafePriceByTimestampOffset => get_lp_tokens_safe_price_by_timestamp_offset
